refatorado componentes
