+++ src/components/HandleFilterXmlByType/index.tsx
