+++ src/components/HeaderTable/index.tsx
@@ -1,0 +1,27 @@
+import React from 'react';
+
+// import { Container } from './styles';
+
+const HeaderTable: any = () => {
+  <thead className="bg-gray-50">
+    <tr>
+      <th className="px-6 py-3 text-xs font-bold text-left text-gray-500 uppercase ">
+        Nnf
+      </th>
+      <th className="px-6 py-3 text-xs font-bold text-left text-gray-500 uppercase ">
+        Chave
+      </th>
+      <th className="px-6 py-3 text-xs font-bold text-left text-gray-500 uppercase ">
+        Data
+      </th>
+      <th className="px-6 py-3 text-xs font-bold text-left text-gray-500 uppercase ">
+        Valor
+      </th>
+      <th className="px-6 py-3 text-xs font-bold text-left text-gray-500 uppercase ">
+        Status
+      </th>
+    </tr>
+  </thead>
+}
+
+export default HeaderTable;
